chapter4/requirements.txt
fastapi
uvicorn[standard]
streamlit
langchain
langchain-google-genai
pydantic
langgraph
chromadb
sentence-transformers>=2.7.0
PyPDF2
tiktoken
requests
python-dotenv
langchain-community
langchain-chroma
langchain-huggingface
chromadb-client
